6.2.3 (4)分页信息数据返回

+++ cloud-api-commons/src/main/java/com/itay/resp/CommonResponse.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Data
 public class CommonResponse<T> {
    List<T> data;
-   Integer total;
-   Integer current;
+   Long total;
+   Long current;
 }
